requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
marshmallow==3.23.0
packaging==24.1
PyJWT==2.9.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.0.6
